주소 수정,comment 분리, 비로그인 상태에서 contact 기능 활성화
