+++ devbox_client/package.json
@@ -2,18 +2,25 @@
   "name": "devbox_client",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:8080",
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.7.7",
+    "bootstrap": "^5.3.3",
+    "devbox_client": "file:",
     "moment": "^2.30.1",
     "react": "^18.3.1",
     "react-calendar": "^5.0.0",
     "react-datepicker": "^7.3.0",
     "react-dom": "^18.3.1",
+    "react-kakao-maps-sdk": "^1.1.27",
     "react-quill": "^2.0.0",
     "react-router-dom": "^6.26.1",
     "react-scripts": "5.0.1",
+    "slick-carousel": "^1.8.1",
+    "styled-components": "^6.1.13",
     "sweetalert2": "^11.13.3",
     "swiper": "^11.1.12",
     "web-vitals": "^2.1.4"
