+++ devbox_client/src/components/Button.jsx
@@ -1,10 +1,11 @@
-const Button = ({ text, onClick }) => {
-    return (
-        <button type="submit" className="btn btn-secondary text-white px-md-4 px-2 py-md-3 py-1 radius-0 light-300"
-            onClick={onClick}>
-            {text}
-        </button>
-    );
+import React from 'react';
+
+const Button = ({ text, onClick, className }) => {
+  return (
+    <button onClick={onClick} className={`btn ${className}`}>
+      {text}
+    </button>
+  );
 };
 
 export default Button;
