+++ devbox_server/src/main/java/com/o2b2/devbox_server/config/CorsMvcConfig.java
@@ -9,9 +9,14 @@ public class CorsMvcConfig implements WebMvcConfigurer {
 
     @Override
     public void addCorsMappings(CorsRegistry corsRegistry) {
-
         corsRegistry.addMapping("/**")
-                .exposedHeaders("Set-Cookie")
-                .allowedOrigins("http://localhost:3000");
+                .allowedOrigins("http://localhost:3000")
+                .allowedMethods("*")
+                .allowedHeaders("*")
+                .allowCredentials(true) // 쿠키를 허용
+                .exposedHeaders("Set-Cookie");
     }
 }
+
+
+
